Move mock_trezorlib et al to conftest in tests.cli
